fix: move model
